rejiggered app.rb to eliminate sinatra views, and move class selection to the server from client
